created and integrated bewerto workflow and workers, refactor workers to re-use more code
