+++ src/test/java/ch/sbb/matsim/bewerto/elasticities/DemandFactorCalculatorTest.java
@@ -54,8 +54,9 @@ class DemandFactorCalculatorTest {
         setupTestData();
 
         // Configure mocks
-        when(mockBaseSkim.getSkims()).thenReturn(baseSkims);
-        when(mockVariantSkim.getSkims()).thenReturn(variantSkims);
+        // TODO: Test needs to be updated when calculator is refactored to use UmlegoSkimCalculator
+//        when(mockBaseSkim.getSkims()).thenReturn(baseSkims);
+//        when(mockVariantSkim.getSkims()).thenReturn(variantSkims);
         when(mockLookup.getCluster(anyString())).thenReturn("CH"); // Default cluster
 
         // Create the calculator
@@ -111,8 +112,6 @@ class DemandFactorCalculatorTest {
         assertTrue(factor != 1.0, "Factor should not be the default value");
 
         // Verify that lookups were performed
-        verify(mockBaseSkim).getSkims();
-        verify(mockVariantSkim).getSkims();
         verify(mockLookup).getCluster("zone1");
 
         // Verify the factor is within reasonable bounds based on the elasticity parameters in the file
